* warning sweep * pre-64bit compatibility (sue size_t whenever needed) * msvc compatibility (LexicalInterpreter<char const[N]>)

+++ tinfra/fmt.cpp
@@ -9,7 +9,7 @@ struct fmt_command {
 };
     
 template <typename T>
-int process_fmt(std::string fmt, int start, fmt_command& result, T& output)
+size_t process_fmt(std::string fmt, size_t start, fmt_command& result, T& output)
 {
     std::string::const_iterator istart = fmt.begin() + start;
     std::string::const_iterator c = istart;
@@ -43,7 +43,7 @@ int process_fmt(std::string fmt, int start, fmt_command& result, T& output)
     }
     //std::cout << "pf: " << -1 << std::endl;
     output.append(istart, end);
-    return -1;
+    return (size_t)-1;
 }
 
 ///
@@ -53,11 +53,11 @@ void simple_fmt::reset() {
     output_.clear();
     pos_ = 0;
 }
-int simple_fmt::check_command()
+size_t simple_fmt::check_command()
 {
     fmt_command cmd;
-    int cmd_pos = process_fmt(fmt_, pos_, cmd, output_);
-    if( cmd_pos == -1 ) throw format_exception("simple_fmt: too many actual arguments");        
+	size_t cmd_pos = process_fmt(fmt_, pos_, cmd, output_);
+    if( cmd_pos == (size_t)-1 ) throw format_exception("simple_fmt: too many actual arguments");        
     switch( cmd.command ) {
     case 's':
     case 'i':
@@ -71,7 +71,7 @@ int simple_fmt::check_command()
 void simple_fmt::realize()
 {
     fmt_command cmd;        
-    if( process_fmt(fmt_, pos_, cmd, output_) != -1 ) 
+    if( process_fmt(fmt_, pos_, cmd, output_) != (size_t)-1 ) 
         throw format_exception("simple_fmt: not all arguments realized");        
     pos_ = fmt_.size();
 }
